Protecting Video Routes

+++ controllers/videoController.js
@@ -29,9 +29,12 @@ export const videoDetail = async(req, res) => {
   const {
     params: { id }
   } = req;
-  const video = await Video.findById(id).populate('creator');
-  console.log(video);
-  res.render('videoDetail', { pageTitle: video.title, video })
+  try {
+    const video = await Video.findById(id).populate('creator');
+    res.render('videoDetail', { pageTitle: video.title, video })
+  } catch (error) {
+    res.redirect(routes.home);
+  }
 }
 
 export const getUpload = (req, res) => {
@@ -49,7 +52,7 @@ export const postUpload = async(req, res) => {
     description,
     creator: req.user.id
   });
-  req.user.videos.push(newVideo.creator);
+  req.user.videos.push(newVideo.id);
   req.user.save();
   res.redirect(routes.videoDetail(newVideo.id));
 }
@@ -60,7 +63,11 @@ export const getEditVideo = async(req, res) => {
   } = req;
   try {
     const video = await Video.findById(id);
-    res.render('editVideo', { pageTitle: `Edit ${video.title}`, video });
+    if (video.creator == req.user.id) {
+      res.render('editVideo', { pageTitle: `Edit ${video.title}`, video });
+    } else {
+      throw Error();
+    }
   } catch (error) {
     res.redirect(routes.home);
   }
@@ -84,7 +91,12 @@ export const deleteVideo = async(req, res) => {
     params: { id }
   } = req;
   try {
-    const video = await Video.findByIdAndRemove({ _id: id });
+    const video = await Video.findById(id);
+    if (video.creator == req.user.id) {
+      await Video.findByIdAndRemove({ _id: id });
+    } else {
+      throw Error;
+    }
   } catch (error) {}
   res.redirect(routes.home);
 }
